Adjusted language to be inclusive.
This switches out the word "crazy" for the more precise "complicated"
since it has historical origins in ableism, and also changes
references to a unit and to an enemy player from he/him to singular
they/them pronouns. These changes are minor but contribute to making
the game as inclusive for women and neurodivergent people as possible
without sacrificing anything.