Fix Enter key not opening chat text input box.
SDL_GetKeyName gives "Return" for the Enter key.

The string "Return" wasn't found in the string table because doesStringExist
does a case-sensitive search for the string and it starts with an uppercase
character, but the string in the string table is all lowercase. Since it
was not found, KeyPress defaults to using the char value of the keysym,
which is '\r'. There isn't an '\r' in the KeyboardManager's list of
shortcuts, so getAction returns DoNothing and the chat box never appears.

This commit also fixes Escape key not opening menu and the Pause button
doing nothing.