Fix confusing if statement in Warrush AI.
I think the if statement tries to check for a building of the same type
and number of workers as the parameters given to the method, and either
under construction or requires workers (e.g. swarm, inn). The five types
of buildings it checks against don't need workers to gather resources
for them.

Fixes #70.